Extract users API endpoints
