+++ app/lib/api/apiRequests.ts
@@ -26,7 +26,7 @@ export interface FileResponse {
     _id: string;
 }
 
-type FilesResult = {
+export type FilesResult = {
     files: FileResponse[];
     totalResults: number;
 };
@@ -69,21 +69,4 @@ export const fetchReplayData = async (file: FileResponse): Promise<ReplayData> =
 
 export const deleteReplay = async (replay: FileResponse) => {
     await apiInstance.delete(`/replays/${replay._id}`);
-};
-
-export type UserInfo = {
-    webId?: string;
-    playerLogin?: string;
-    playerName?: number;
-    _id?: number;
-};
-
-export const getUserInfo = async (webId: string): Promise<UserInfo> => {
-    const res = await apiInstance.get(`/users/${webId}/info`);
-    return res.data;
-};
-
-export const getUserReplays = async (webId: string): Promise<FilesResult> => {
-    const res = await apiInstance.get(`/users/${webId}/replays`);
-    return res.data;
 };
