+++ app/lib/api/apiWrapper.ts
@@ -1,7 +1,9 @@
 import * as maps from './requests/maps';
+import * as users from './requests/users';
 
 const api = {
     maps,
+    users,
 };
 
 export default api;
